requirements.txt
audioread==2.1.6
cycler==0.10.0
decorator==4.3.0
ffmpeg-python==0.1.17
future==0.17.1
joblib==0.13.0
kiwisolver==1.0.1
librosa==0.6.2
llvmlite==0.26.0
matplotlib==3.0.2
numba==0.41.0
numpy==1.15.4
opencv-python==3.4.5.20
pyparsing==2.3.0
python-dateutil==2.7.5
resampy==0.2.1
scikit-learn==0.20.2
scipy==1.2.0
six==1.12.0
youtube-dl==2018.12.17
